API/requirements.txt
Django==1.8
djangorestframework==3.3.3
pkg-resources==0.0.0
